make gift timeout an int instead of byte, allowing for much larger timeouts

+++ crawl-ref/source/player.h
@@ -228,7 +228,7 @@ public:
     string jiyva_second_name;       // Random second name of Jiyva
     uint8_t piety;
     uint8_t piety_hysteresis;       // amount of stored-up docking
-    uint8_t gift_timeout;
+    int gift_timeout;
     uint8_t saved_good_god_piety;   // for if you "switch" between E/Z/1 by abandoning one first
     god_type previous_good_god;
     FixedVector<uint8_t, NUM_GODS>  penance;
